+++ mods/gen1_wild_ui_nightly/CHANGELOG.md
@@ -6,6 +6,12 @@ was taken from.
 
 [stable]: https://github.com/wild1walker/Gen1WildUI
 
+## [0.32.59] - 2026-09-05
+
+### Changed
+
+- Version moved with the nightly channel; nothing in this bundle changed.
+
 ## [0.32.58] - 2026-09-05
 
 ### Fixed